requirements.txt
streamlit==1.32.0
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
python-jose==3.3.0
requests==2.31.0
python-multipart==0.0.9
python-dateutil==2.8.2 